+++ sdks/go/data/git/git.go
@@ -4,29 +4,29 @@ package git
 
 import (
 	"context"
+	"crypto/sha1"
+	"fmt"
+	"os"
 	"path/filepath"
 
-	"github.com/opctl/opctl/sdks/go/data/fs"
 	"github.com/opctl/opctl/sdks/go/model"
-	"github.com/pkg/errors"
 	"golang.org/x/sync/singleflight"
 )
 
-// New returns a data provider which sources pkgs from git repos
-func New(basePath string) model.DataProvider {
+// singleFlightGroup is used to ensure resolves don't race across provider intances
+var resolveSingleFlightGroup singleflight.Group
+
+// New returns a data provider which sources data from git repos
+func New(
+	basePath string,
+) model.DataProvider {
 	return &_git{
-		localFSProvider: fs.New(basePath),
-		basePath:        basePath,
+		basePath: basePath,
 	}
 }
 
 type _git struct {
-	// composed of fsProvider
-	localFSProvider model.DataProvider
-	basePath        string
-
-	// resolveSingleFlightGroup is used to ensure resolves don't race across provider instances
-	resolveSingleFlightGroup singleflight.Group
+	basePath string
 }
 
 func (gp *_git) Label() string {
@@ -40,26 +40,36 @@ func (gp *_git) TryResolve(
 	dataRef string,
 ) (model.DataHandle, error) {
 	// attempt to resolve within singleFlight.Group to ensure concurrent resolves don't race
-	handle, err, _ := gp.resolveSingleFlightGroup.Do(
+	handle, err, _ := resolveSingleFlightGroup.Do(
 		dataRef,
 		func() (interface{}, error) {
-			parsedPkgRef, err := parseRef(dataRef)
-			if err != nil {
-				return nil, errors.Wrap(err, "invalid git ref")
-			}
+			repoPath := filepath.Join(gp.basePath, dataRef)
+			handle := newHandle(repoPath, dataRef)
 
-			// attempt to resolve from cache
-			handle, _ := gp.localFSProvider.TryResolve(ctx, eventChannel, callID, dataRef)
-			// ignore errors from local resolution, since we'll try to pull from a remote
-			if handle != nil {
+			// we'll mark complete clones in case we get interrupted
+			completeMarkerPath := filepath.Join(gp.basePath, fmt.Sprintf(".%x", sha1.Sum([]byte(dataRef))))
+
+			_, err := os.Stat(completeMarkerPath)
+			if err == nil {
+				// complete clone found
 				return handle, nil
+			} else if !os.IsNotExist(err) {
+				// incomplete clone; blow it away
+				if err := os.RemoveAll(repoPath); err != nil {
+					return nil, err
+				}
 			}
 
-			// attempt pull if cache miss
-			if err := gp.pull(ctx, eventChannel, callID, parsedPkgRef); err != nil {
+			// attempt clone
+			if err := gp.Clone(ctx, eventChannel, callID, dataRef); err != nil {
 				return nil, err
 			}
-			return newHandle(filepath.Join(gp.basePath, dataRef), dataRef), nil
+
+			// mark complete
+			f, err := os.Create(completeMarkerPath)
+			f.Close()
+
+			return handle, err
 		},
 	)
 	if err != nil {